+++ src/ast/mod.rs
@@ -916,6 +916,9 @@ fn normalize_part_usage_body_element_node(
         PartUsageBodyElement::ConstraintUsage(n) => {
             PartUsageBodyElement::ConstraintUsage(dummy_node(n, n.value.clone()))
         }
+        PartUsageBodyElement::CalcUsage(n) => {
+            PartUsageBodyElement::CalcUsage(dummy_node(n, n.value.clone()))
+        }
         PartUsageBodyElement::Import(n) => {
             PartUsageBodyElement::Import(dummy_node(n, normalize_import(&n.value)))
         }